Group element visibility implementation
